+++ R/redcap.R
@@ -660,6 +660,7 @@ select_rc_classic <- function(rc_data, var_name, metadata, checkbox_aux) {
 #'
 #' @param rc_data RedCap data imported with ody_rc_import.
 #' @param ... Variable names to select. If the name of a form is provided, all the variables belonguing to that form will be selected.
+#' @param .is_vector Logical. If TRUE, the first element of ... is considered a character vector with the names of the variables to be selected.
 #' @param .if_different_forms What action take if the selected variables belong to different forms.
 #'    - list: It returns a list with an element for each form so only variables belonging to the same form are joinned in the same data frame.
 #'    - join: Join all variables creating artifact NAs.
@@ -669,6 +670,7 @@ select_rc_classic <- function(rc_data, var_name, metadata, checkbox_aux) {
 #' @export
 ody_rc_select <- function(rc_data,
                           ...,
+                          .is_vector = FALSE,
                           .if_different_forms = c("list", "join"),
                           .include_aux = FALSE) {
 
@@ -680,6 +682,11 @@ ody_rc_select <- function(rc_data,
   ) |> purrr::map(as.character) |>
     purrr::reduce(c) |>
     unique()
+
+  # trick to permit character vectors
+  if (.is_vector) {
+      sel_vars <- get(sel_vars[1])
+  }
 
   if (names(rc_data)[1] == "redcap_event_name") {
     select_rc_function <- select_rc_long
@@ -744,7 +751,103 @@ ody_rc_select <- function(rc_data,
         sel_vars,
         function(x) select_rc_function(rc_data, x, metadata, checkbox_aux)
       ),
-      form = purrr::map_chr(
+      form = purrr::map(
+        .data$variables, ~.$redcap_form_name |> unique()
+      )
+    ) |>
+      tidyr::nest(data = .data$variables)
+
+    extracted_list <- purrr::map(
+      extracted_vars[[2]],
+      ~purrr::reduce(.[[1]], dplyr::full_join)
+    ) |> suppressMessages()
+
+
+    names(extracted_list) <- extracted_vars$form
+
+    if (length(extracted_list) == 1) {
+      extracted_list[[1]]
+    } else {
+      extracted_list
+    }
+
+  }
+
+}
+
+# Helper function used in viewer apps that directly assume variable is
+# a vector
+rc_select_viewer <- function(rc_data,
+                             sel_vars,
+                             .if_different_forms = c("list", "join"),
+                             .include_aux = FALSE) {
+
+  .if_different_forms <- rlang::arg_match(.if_different_forms)
+
+  if (names(rc_data)[1] == "redcap_event_name") {
+    select_rc_function <- select_rc_long
+  } else {
+    select_rc_function <- select_rc_classic
+  }
+
+  metadata <- attr(rc_data, "metadata")
+  checkbox_aux <- attr(rc_data, "checkbox_aux")
+
+  # If a form name is provided, all the variables of the form are extracted
+  if (length(sel_vars) == 1 && sel_vars %in% unique(metadata$form_name)) {
+    current_form <- metadata |>
+      dplyr::filter(
+        .data[["form_name"]] == sel_vars
+      ) |>
+      dplyr::pull("form_name") |>
+      unique()
+
+    # If the form contains phantom variables, the must be excluded since they do
+    # not actually exist and the selection function would fail.
+    phantom_vars <- attr(rc_data, "phantom_variables") |>
+      dplyr::pull("field_name")
+
+    sel_vars <- metadata |>
+      dplyr::filter(
+        .data[["form_name"]] == current_form,
+        !(.data[["field_name"]] %in% phantom_vars)
+      ) |>
+      dplyr::pull("field_name")
+
+    if (.include_aux) {
+      checkbox_vars <- metadata |>
+        dplyr::filter(
+          .data$field_name %in% sel_vars,
+          .data$field_type == "checkbox"
+        ) |>
+        dplyr::pull("field_name") |>
+        stringr::str_c(collapse = "|")
+
+      needed_aux <- stringr::str_subset(checkbox_aux, checkbox_vars)
+
+      sel_vars <- c(sel_vars, needed_aux)
+
+    }
+
+  }
+
+  if (.if_different_forms == "join") {
+
+    purrr::map(
+      sel_vars,
+      function(x) select_rc_function(rc_data, x, metadata, checkbox_aux)
+    ) |>
+      purrr::reduce(dplyr::full_join) |>
+      suppressMessages()
+
+  } else {
+
+    extracted_vars <- tibble::tibble(
+      variables = purrr::map(
+        sel_vars,
+        function(x) select_rc_function(rc_data, x, metadata, checkbox_aux)
+      ),
+      form = purrr::map(
         .data$variables, ~.$redcap_form_name |> unique()
       )
     ) |>
@@ -1186,5 +1289,22 @@ ody_rc_spread <- function(rc_data = NULL) {
     spread_repeating |>
       dplyr::select(tidyselect::starts_with(fields))
   }
+
+}
+
+#' Add the redcap_data import date to a file name
+#'
+#' @param file_name The name of the file to be saved.
+#' @param extension The extension of the file. Default is "csv".
+#'
+#' @return A string with the new file name.
+#' @export
+ody_rc_add_import_date <- function(file_name, extension = "csv") {
+
+  loaded_date <- ody_rc_current(as_list = TRUE)$loaded |>
+    stringr::str_remove_all("-|:") |>
+    stringr::str_replace_all(" ", "_")
+
+  stringr::str_c(file_name, "_", loaded_date, ".", extension)
 
 }